# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ITBF API",
        "description": "API for ITBF",
        "contact": {
            "name": "Christian Lerma",
            "url": "",
            "email": "christianlerma@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/hoteles": {
            "get": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Get all registers",
                "operationId": "20fe4f64d792f6f07492e53657a77000",
                "responses": {
                    "200": {
                        "description": "Retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Create a new register",
                "operationId": "bdc7e055f7388a1048bb054703909745",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hotel"
                                ],
                                "properties": {
                                    "hotel": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "direccion": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "pagina_web": {
                                        "type": "string"
                                    },
                                    "calificacion": {
                                        "type": "integer"
                                    },
                                    "numero_habitaciones": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/hoteles/{id}": {
            "get": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Get a register by ID",
                "operationId": "932399851018abdc74f3f3cd2c44d09a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/hoteles/{id}/editar": {
            "put": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Update a register",
                "operationId": "59fceca8cbc96cebe32c01a161f0e2d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the register",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hotel"
                                ],
                                "properties": {
                                    "hotel": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "direccion": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "pagina_web": {
                                        "type": "string"
                                    },
                                    "calificacion": {
                                        "type": "integer"
                                    },
                                    "numero_habitaciones": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error updating the register",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/hoteles/{id}/eliminar": {
            "delete": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Delete a register",
                "operationId": "e5366ad5d13e5b7f10626f6d33a6a916",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}